+++ src/oauth/tokenStorage/serverCookie.ts
@@ -2,6 +2,7 @@ import { serialize, parse } from "cookie"
 
 import type {OAuthToken} from "../../oauth"
 import type {TokenStorage} from "../tokenStorage"
+import {cookieName} from "../tokenStorage"
 
 let requestCookies = ""
 let responseCookie = ""
@@ -16,24 +17,24 @@ export const getResponseCookie = (): string => {
 
 export const serverCookie: TokenStorage = {
     remove() {
-        responseCookie = serialize("svelte-oauth-token", "deleted", {
+        responseCookie = serialize(cookieName, "deleted", {
             expires: new Date(0),
             sameSite: "strict"
         })
     },
     get(): OAuthToken | null | undefined {
         const cookies = parse(requestCookies)
-        if (!Object.keys(cookies).includes("svelte-oauth-token")) {
+        if (!Object.keys(cookies).includes(cookieName)) {
             return null
         }
 
         try {
-            return JSON.parse(cookies["svelte-oauth-token"])
+            return JSON.parse(cookies[cookieName])
         } catch (e) {
             return null
         }
     }, set(token: OAuthToken): void {
-        responseCookie = serialize("svelte-oauth-token", JSON.stringify(token), {
+        responseCookie = serialize(cookieName, JSON.stringify(token), {
             sameSite: "strict"
         })
         requestCookies = responseCookie