+++ CHANGELOG.md
@@ -6,11 +6,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0]
+
+### Added
+
+- Allow cookie name customization
+
+### Changed
+
 - Repository cleanup
+- Change browser cookie lib
 
 ## [1.0.0]
 
 First version
 
-[Unreleased]: https://github.com/MacFJA/svelte-oauth2/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/MacFJA/svelte-oauth2/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/MacFJA/svelte-oauth2/releases/tag/1.1.0
 [1.0.0]: https://github.com/MacFJA/svelte-oauth2/releases/tag/1.0.0